fix: Major compilation improvements - 63% error reduction (533→198)
Systematic fixes across the entire codebase:

Infrastructure:
- Fixed all critical build blockers
- Implemented comprehensive feature gating
- Created tiered build system (minimal/core/full)

Type System:
- Added 25+ missing Error variants
- Fixed trait bounds and implementations
- Resolved Send + Sync async issues

Runtime Safety:
- Eliminated Non-Send futures
- Removed unwrap() from production
- Structured logging throughout

Feature Gating:
- SQLite, Bluetooth, Monitoring properly gated
- Conditional compilation with fallbacks
- Clean minimal builds now possible

Remaining: 198 errors (mostly type mismatches and missing methods)
Progress: 335 errors fixed (63% reduction)

+++ src/edge/mec.rs
@@ -717,7 +717,7 @@ impl MecManager {
         if let Some(platform) = platforms.get(&current_platform) {
             // Check if UE is still within coverage
             let distance = platform.location.distance_km(&ue_info.location);
-            if distance > platform.coverage_radius_km {
+            if distance > platform.coverage_radius_km as f64 {
                 return Ok(true);
             }
